+++ src/lib.rs
@@ -1,10 +1,10 @@
 pub mod add;
+pub mod check;
 pub mod config;
 pub mod file_picker;
 pub mod misc;
 pub mod upgrade;
 
-// Get the home directory when the program starts
 lazy_static::lazy_static! {
     pub static ref HOME: std::path::PathBuf = home::home_dir().expect("Could not get user's home directory");
 }